fix some warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4477 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cpu/tms7000/tms7000.c
@@ -512,7 +512,7 @@ UINT8 tms7000_calculate_timer1_decrementator( void )
 	double prescalertimer = TIME_IN_CYCLES(16,cpuNum) *
 													(tms7000.pf[0x03] & 0x1f);
 	
-	result = (tms7000.time_timer1 - timer_get_time()) / prescalertimer;
+	result = (UINT8)((tms7000.time_timer1 - timer_get_time()) / prescalertimer);
 
 	return result;
 }
